refactor: 패키지 이동
관련된 dto 들 끼리 묶었고,
기존에 있던 StationRepository 도 도메인 외부로 이동시켰다.
